streamlit_frontend/requirements.txt
streamlit>=1.28.0
pandas>=2.1.0
numpy>=1.24.3
matplotlib>=3.7.2
altair>=5.0.1
pillow>=10.0.0
python-dotenv>=1.0.0
requests>=2.31.0
PyPDF2>=3.0.1
pdfplumber>=0.10.2
pdfminer.six>=20221105
python-docx>=0.8.11
spacy>=3.6.1
nltk>=3.8.1
# Run 'python -m spacy download en_core_web_sm' after installing dependencies
python-multipart>=0.0.6
scikit-learn>=1.3.0
# PDF and document processing
pdfplumber>=0.10.2
PyPDF2>=3.0.1
python-docx>=0.8.11
# OCR and image processing
pdf2image>=1.16.3
pytesseract>=0.3.10
opencv-python>=4.8.0.76
Pillow>=10.0.0
# Optional: For enhanced visualization
plotly>=5.16.1
wordcloud>=1.9.2
# For testing and development
reportlab>=4.0.4
streamlit-authenticator>=0.3.0
streamlit-cropper 